cleaned unecessary props in CustomTag
